[tests] Refactored test infrastructure - Split into generic and target specific tests - Split subtests into where they are executed (host, snitchCluster)
